requirements.txt
Django==2.2.2
django-taggit==1.1.0
mysqlclient==1.4.2.post1
django-tinymce==2.8.0
